(Server): Geometry对象导出WKT；读取WKT创建Geometry对象

+++ Server/src/geom/coord.go
@@ -2,13 +2,13 @@ package geom
 
 // Coord is the basic type of gemotry with 2-D position
 type Coord struct {
-	x float32
-	y float32
+	x float64
+	y float64
 }
 
 // Coord is the basic type of gemotry with 3-D position
 type Coord3D struct {
-	x float32
-	y float32
-	z float32
+	x float64
+	y float64
+	z float64
 }
